requirements.txt
redis
msgpack-python
pytest
